+++ include/helpers/PlatformHelper.h
