+++ src/helpers/PlatformHelper.cpp
