+++ src/managers/ConnManager.cpp
